requirements.txt
django==3.2.4
colorgram.py==1.2.0
fiftyone==0.9.3